backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["iqaz100 <piotrek10.lasek@o2.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
jwt = "^1.3.1"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
